Look up sensor names between run_melb() and walk_melb()

One-to-one corresponding sensor names between run_melb() and walk_melb()

lookup_sensor()

Value

A data frame including three columns:

run: Sensor names obtained from the run_melb() function using Socrata

walk: Sensor names obtained from the walk_melb() function using compedapi

match: whether sensor names are identical or not

Details

Two APIs (Socrata and compedapi) code some sensors using different names. This functions returns a data frame that allows to compare sensor names obtained from these two APIs.

Examples

  lookup_sensor()

#>                                       run                            walk match
#> 1                           State Library                   State Library  TRUE
#> 2                   Collins Place (South)           Collins Place (South)  TRUE
#> 3                   Collins Place (North)           Collins Place (North)  TRUE
#> 4                       Flagstaff Station               Flagstaff Station  TRUE
#> 5                       Melbourne Central               Melbourne Central  TRUE
#> 6                        Town Hall (West)                Town Hall (West)  TRUE
#> 7              Bourke Street Mall (North)      Bourke Street Mall (North)  TRUE
#> 8              Bourke Street Mall (South)      Bourke Street Mall (South)  TRUE
#> 9                    Australia on Collins            Australia on Collins  TRUE
#> 10                 Southern Cross Station          Southern Cross Station  TRUE
#> 11                         Victoria Point                  Victoria Point  TRUE
#> 12                               New Quay                        New Quay  TRUE
#> 13                        Waterfront City                 Waterfront City  TRUE
#> 14                            Webb Bridge                     Webb Bridge  TRUE
#> 15                         Princes Bridge                  Princes Bridge  TRUE
#> 16                       Sandridge Bridge                Sandridge Bridge  TRUE
#> 17                         Birrarung Marr                  Birrarung Marr  TRUE
#> 18        Flinders St-Elizabeth St (East) Flinders St-Elizabeth St (East)  TRUE
#> 19          Spencer St-Collins St (North)   Spencer St-Collins St (North)  TRUE
#> 20          Spencer St-Collins St (South)   Spencer St-Collins St (South)  TRUE
#> 21            Bourke St-Russell St (West)     Bourke St-Russell St (West)  TRUE
#> 22          Chinatown-Swanston St (North)   Chinatown-Swanston St (North)  TRUE
#> 23         Chinatown-Lt Bourke St (South)  Chinatown-Lt Bourke St (South)  TRUE
#> 24                      QV Market-Peel St               QV Market-Peel St  TRUE
#> 25                    Lonsdale St (South)             Lonsdale St (South)  TRUE
#> 26                        Lygon St (West)                 Lygon St (West)  TRUE
#> 27           Flinders St-Spring St (West)    Flinders St-Spring St (West)  TRUE
#> 28                           Alfred Place                    Alfred Place  TRUE
#> 29         Flinders St-Swanston St (West)  Flinders St-Swanston St (West)  TRUE
#> 30                            City Square                     City Square  TRUE
#> 31          Grattan St-Swanston St (West)   Grattan St-Swanston St (West)  TRUE
#> 32           Monash Rd-Swanston St (West)    Monash Rd-Swanston St (West)  TRUE
#> 33           Tin Alley-Swanston St (West)    Tin Alley-Swanston St (West)  TRUE
#> 34                              Southbank                       Southbank  TRUE
#> 35      Flinders Street Station Underpass   Flinders St Station Underpass FALSE
#> 36          QV Market-Elizabeth St (West)      QV Market-Elizabeth (West) FALSE
#> 37 Melbourne Convention Exhibition Centre    Convention/Exhibition Centre FALSE
#> 38                        The Arts Centre                 Vic Arts Centre FALSE
#> 39                   Flinders St-Spark La          Flinders St-Spark Lane FALSE
#> 40                        Queen St (West)             Queen Street (West) FALSE
#> 41                        Lygon St (East)             Lygon Street (East) FALSE
#> 42           Lonsdale St-Spring St (West)   Spring St-Lonsdale St (South) FALSE
#> 43          St Kilda Rd-Alexandra Gardens     St. Kilda-Alexandra Gardens FALSE
#> 44         Flinders La-Swanston St (West)                              NA FALSE
#> 45   Little Collins St-Swanston St (East)                              NA FALSE