API using compedapi to Melbourne pedestrian data

Provides API using compedapi to Melbourne pedestrian data in a tidy data form.

melb_walk(from = to - 6L, to = Sys.Date() - 1L, tz = "",
  na.rm = FALSE, session = NULL)

Arguments

from
Starting date.

to
Ending date.

tz
Deprecated. For this dataset, it should only be "Australia/Melbourne".

na.rm
Logical. FALSE is the default suggesting to include NA in the dataset. TRUE removes the NAs.

session
NULL or "shiny". For internal use only.

Value

A tibble including these variables as follows:

Sensor: Sensor name (43 sensors up to date)

Date_Time: Date time when the pedestrian counts are recorded

Date: Date associated with Date_Time

Time: Time of day

Count: Hourly counts

Details

It provides API using compedapi, where counts are uploaded on a daily basis. The up-to-date data would be till the previous day. The data is sourced from Melbourne Open Data Portal. Please refer to Melbourne Open Data Portal for more details about the dataset and its policy.

See also

Examples

# Retrieve last week data
melb_walk()

#> # A tibble: 7,224 x 5
#>    Sensor                            Date_Time           Date        Time Count
#>    <chr>                             <dttm>              <date>     <int> <int>
#>  1 Bourke Street Mall (North)        2018-12-06 00:00:00 2018-12-06     0   133
#>  2 Bourke Street Mall (South)        2018-12-06 00:00:00 2018-12-06     0    76
#>  3 Melbourne Central                 2018-12-06 00:00:00 2018-12-06     0   303
#>  4 Town Hall (West)                  2018-12-06 00:00:00 2018-12-06     0   235
#>  5 Princes Bridge                    2018-12-06 00:00:00 2018-12-06     0   194
#>  6 Flinders Street Station Underpass 2018-12-06 00:00:00 2018-12-06     0   293
#>  7 Birrarung Marr                    2018-12-06 00:00:00 2018-12-06     0    45
#>  8 Webb Bridge                       2018-12-06 00:00:00 2018-12-06     0    25
#>  9 Southern Cross Station            2018-12-06 00:00:00 2018-12-06     0    33
#> 10 Victoria Point                    2018-12-06 00:00:00 2018-12-06     0     8
#> # … with 7,214 more rows

# Retrieve data of a speficied period
start_date <- as.Date("2017-07-01")
end_date <- start_date + 6L
melb_walk(from = start_date, to = end_date)

#> # A tibble: 7,224 x 5
#>    Sensor                            Date_Time           Date        Time Count
#>    <chr>                             <dttm>              <date>     <int> <int>
#>  1 Bourke Street Mall (North)        2017-07-01 00:00:00 2017-07-01     0   280
#>  2 Bourke Street Mall (South)        2017-07-01 00:00:00 2017-07-01     0   177
#>  3 Melbourne Central                 2017-07-01 00:00:00 2017-07-01     0   826
#>  4 Town Hall (West)                  2017-07-01 00:00:00 2017-07-01     0   682
#>  5 Princes Bridge                    2017-07-01 00:00:00 2017-07-01     0     0
#>  6 Flinders Street Station Underpass 2017-07-01 00:00:00 2017-07-01     0   768
#>  7 Birrarung Marr                    2017-07-01 00:00:00 2017-07-01     0    74
#>  8 Webb Bridge                       2017-07-01 00:00:00 2017-07-01     0    22
#>  9 Southern Cross Station            2017-07-01 00:00:00 2017-07-01     0    35
#> 10 Victoria Point                    2017-07-01 00:00:00 2017-07-01     0    27
#> # … with 7,214 more rows