API using Socrata to Melbourne pedestrian data with directions (per minute)

API using Socrata to Melbourne pedestrian data with directions (per minute)

melb_walk_directional(app_token = NULL)

Arguments

app_token
Characters giving the application token. A limited number of requests can be made without an app token (NULL), but they are subject to much lower throttling limits than request that do include one. Sign up for an app token here.

Value

A tibble including these variables as follows:

sensor_id: Sensor name

date_time: Date time when the pedestrian counts are recorded

date: Date associated with date_time

time: Time of day

direction_1: Direction 1 sensor reading (count of pedestrians)

direction_2: Direction 2 sensor reading (count of pedestrians)

total_of_directions: Total sensor reading i.e. direction 1+2 (count of pedestrians)

Details

It provides the API using Socrata, to access minute by minute directional pedestrian counts for the last hour from pedestrian sensor devices located across the city. The data is updated every 15 minutes.

Columns sensor_id, direction_1, and direction_2 can be used to join the data with the Sensor Locations dataset which details the location, status, and directional readings of sensors, which can be obtained from pull_sensor().

See also

Examples

# \dontrun{
melb_walk_directional()

#> # A tibble: 1,759 x 7
#>    sensor_id date_time           date       time  direction_1 direction_2
#>    <chr>     <dttm>              <date>     <tim>       <int>       <int>
#>  1 47        2020-03-21 06:56:00 2020-03-21 06:56           1           1
#>  2 12        2020-03-21 06:56:00 2020-03-21 06:56           1           0
#>  3 3         2020-03-21 06:56:00 2020-03-21 06:56           0           4
#>  4 53        2020-03-21 06:56:00 2020-03-21 06:56           0           3
#>  5 4         2020-03-21 06:56:00 2020-03-21 06:56           1           2
#>  6 56        2020-03-21 06:56:00 2020-03-21 06:56           1           0
#>  7 31        2020-03-21 06:56:00 2020-03-21 06:56           1           1
#>  8 58        2020-03-21 06:56:00 2020-03-21 06:56           1           0
#>  9 39        2020-03-21 06:56:00 2020-03-21 06:56           0           1
#> 10 28        2020-03-21 06:56:00 2020-03-21 06:56           2           0
#> # … with 1,749 more rows, and 1 more variable: total_of_directions <int>

# }