API to access Melbourne microclimate sensor data

API to access Melbourne microclimate sensor data

melb_weather(
  from = to - 6L,
  to = Sys.Date(),
  site = NULL,
  sensor_type = NULL,
  app_token = NULL
)

Arguments

from
Starting date. Earliest measurement is 2019-11-15

to
Ending date.

site
The site identifier. By default will pull in all locations that have weather sensors pull_weather_sensors().

sensor_type
The type of microclimate measurement to extract see pull_weather_types() for details.

app_token
Characters giving the application token. A limited number of requests can be made without an app token (NULL), but they are subject to much lower throttling limits than request that do include one. Sign up for an app token here.

Value

A tibble including these variables as follows:

site: Site identifier, this is the location of the weather sensor

date_time: Date time when the measurement was recorded

date: Date associated with date_time

sensor_type: The type of microclimate sensor reading

units: The units that value is in

value: The value of the reading

Details

It provides the API using Socrata, where microclimate measurements are updated on a dailly basis. For data documentation, including a data dictionary see the Melbourne Open Data Portal. Please refer to Melbourne Open Data Portal for more details about the dataset and its policy.

See also

Examples

if (FALSE) {
# Retrieve the last weeks data
melb_weather()

# Retrieve the last week but for a single location (Pelham St, Carlton)
melb_weather(site = "arc1047")

# Retrieve the last week but only ambient air temperature
melb_weather(sensor_type = "TPH.TEMP")

}