API using Socrata to Melbourne pedestrian sensor locations

Provides API using Socrata to Melbourne pedestrian sensor locations.

pull_sensor(app_token = NULL)

Arguments

app_token
Characters giving the application token. A limited number of requests can be made without an app token (NULL), but they are subject to much lower throttling limits than request that do include one. Sign up for an app token here.

Value

A data frame including these variables as follows:

Sensor: Sensor name (43 sensors up to date)

Sensor_ID: Sensor identifier

Longitude: Longitude

Latitude: Latitude

Location_Type: Location type

Year_Installed: Year installed

Details

It provides API using Socrata.

See also

Examples

pull_sensor()

#>                                                                        Sensor
#> 1                                                           Melbourne Central
#> 2  Melbourne Central-Elizabeth St (East)Melbourne Central-Elizabeth St (East)
#> 3                                                         QVM-Queen St (East)
#> 4                                                           QV Market-Peel St
#> 5                                                             Lygon St (West)
#> 6                                                     QVM-Franklin St (North)
#> 7                                                     Lincoln-Swanston (West)
#> 8                                                         Lonsdale St (South)
#> 9                                                      Collins Street (North)
#> 10                                                          Flagstaff Station
#> 11                                       Little Collins St-Swanston St (East)
#> 12                                              QV Market-Elizabeth St (West)
#> 13                                               Lonsdale St-Spring St (West)
#> 14                                                 Faraday St-Lygon St (West)
#> 15                                                                Webb Bridge
#> 16                                                           Sandridge Bridge
#> 17                                             Flinders La-Swanston St (West)
#> 18                                                      Collins Place (North)
#> 19                                            Elizabeth St-La Trobe St (East)
#> 20                                                             Birrarung Marr
#> 21                                                 Bourke Street Mall (South)
#> 22                                             Bourke St - Spencer St (North)
#> 23                                                           Bourke St Bridge
#> 24                                              Spencer St-Collins St (South)
#> 25                                                            Lygon St (East)
#> 26                                          Flinders Street Station Underpass
#> 27                                                       Australia on Collins
#> 28                                                            The Arts Centre
#> 29                                                      Collins Place (South)
#> 30                                                               Alfred Place
#> 31                                                       Flinders St-Spark La
#> 32                                                     Southern Cross Station
#> 33                                         Lonsdale St - Elizabeth St (North)
#> 34                                                            Waterfront City
#> 35                                                      QVM-Therry St (South)
#> 36                                              St Kilda Rd-Alexandra Gardens
#> 37                                                             Victoria Point
#> 38                                                                   New Quay
#> 39                                               Tin Alley-Swanston St (West)
#> 40                                               Monash Rd-Swanston St (West)
#> 41                                                          Pelham St (South)
#> 42                                                           Town Hall (West)
#> 43                                              Grattan St-Swanston St (West)
#> 44                                                             Princes Bridge
#> 45                                                 Bourke Street Mall (North)
#> 46                                     Melbourne Convention Exhibition Centre
#> 47                                                                  Southbank
#> 48                                            Flinders St-Elizabeth St (East)
#> 49                                                Bourke St-Russell St (West)
#> 50                                           Elizabeth St-Lonsdale St (South)
#>    Sensor_ID Longitude  Latitude Location_Type Year_Installed
#> 1          3  144.9643 -37.81102       Outdoor           2009
#> 2         47  144.9622 -37.81160       Outdoor           2017
#> 3         48  144.9587 -37.80632       Outdoor           2017
#> 4         27  144.9564 -37.80607       Outdoor           2013
#> 5         31  144.9666 -37.80170       Outdoor           2014
#> 6         51  144.9591 -37.80842       Outdoor           2017
#> 7         54  144.9631 -37.80402       Outdoor           2018
#> 8         30  144.9666 -37.81122       Outdoor           2014
#> 9         53  144.9655 -37.81564       Outdoor           2015
#> 10        13  144.9565 -37.81240       Outdoor           2009
#> 11        45  144.9661 -37.81414       Outdoor           2017
#> 12        26  144.9605 -37.80870       Outdoor           2013
#> 13        40  144.9723 -37.80999       Outdoor           2015
#> 14        50  144.9672 -37.79808       Outdoor           2017
#> 15         8  144.9472 -37.82294       Outdoor           2009
#> 16        14  144.9629 -37.82011       Outdoor           2009
#> 17        41  144.9669 -37.81669       Outdoor           2017
#> 18        18  144.9731 -37.81345       Outdoor           2009
#> 19        55  144.9614 -37.81006       Outdoor           2018
#> 20         7  144.9714 -37.81832       Outdoor           2009
#> 21         2  144.9652 -37.81381       Outdoor           2009
#> 22        58  144.9536 -37.81686       Outdoor           2018
#> 23        57  144.9503 -37.81767       Outdoor           2018
#> 24        23  144.9545 -37.81909       Outdoor           2013
#> 25        37  144.9667 -37.80310       Outdoor           2015
#> 26         6  144.9656 -37.81912       Outdoor           2009
#> 27        16  144.9652 -37.81573       Outdoor           2009
#> 28        28  144.9688 -37.82130       Outdoor           2013
#> 29        17  144.9732 -37.81363       Outdoor           2009
#> 30        39  144.9700 -37.81380       Outdoor           2015
#> 31        34  144.9742 -37.81538       Outdoor           2014
#> 32         9  144.9510 -37.81983       Outdoor           2009
#> 33        56  144.9615 -37.81235       Outdoor           2018
#> 34        11  144.9397 -37.81565       Outdoor           2009
#> 35        49  144.9596 -37.80730       Outdoor           2017
#> 36        29  144.9686 -37.81966       Outdoor           2014
#> 37        10  144.9471 -37.81876       Outdoor           2009
#> 38        12  144.9429 -37.81458       Outdoor           2009
#> 39        44  144.9644 -37.79690       Outdoor           2015
#> 40        43  144.9641 -37.79845       Outdoor           2015
#> 41        46  144.9616 -37.80241       Outdoor           2017
#> 42         4  144.9661 -37.81488       Outdoor           2009
#> 43        42  144.9639 -37.80009       Outdoor           2015
#> 44         5  144.9679 -37.81874       Outdoor           2009
#> 45         1  144.9652 -37.81349       Outdoor           2009
#> 46        25  144.9560 -37.82402       Outdoor           2013
#> 47        35  144.9651 -37.82018       Outdoor           2016
#> 48        22  144.9651 -37.81786       Outdoor           2013
#> 49        21  144.9678 -37.81245       Outdoor           2013
#> 50        52  144.9619 -37.81252       Outdoor           2017