API using compedapi to Melbourne pedestrian data

Provides API using compedapi to Melbourne pedestrian data in a tidy data form.

melb_walk(from = to - 6L, to = Sys.Date() - 1L, na.rm = FALSE,
  session = NULL)

Arguments

from
Starting date.

to
Ending date.

na.rm
Logical. FALSE is the default suggesting to include NA in the dataset. TRUE removes the NAs.

session
NULL or "shiny". For internal use only.

Value

A tibble including these variables as follows:

Sensor: Sensor name (43 sensors up to date)

Date_Time: Date time when the pedestrian counts are recorded

Date: Date associated with Date_Time

Time: Time of day

Count: Hourly counts

Details

It provides API using compedapi, where counts are uploaded on a daily basis. The up-to-date data would be till the previous day. The data is sourced from Melbourne Open Data Portal. Please refer to Melbourne Open Data Portal for more details about the dataset and its policy.

See also

Examples

# \dontrun{
# Retrieve last week data
melb_walk()

#> # A tibble: 7,224 x 5
#>    Sensor                            Date_Time           Date        Time Count
#>    <chr>                             <dttm>              <date>     <int> <int>
#>  1 Bourke Street Mall (North)        2019-11-11 00:00:00 2019-11-11     0    70
#>  2 Bourke Street Mall (South)        2019-11-11 00:00:00 2019-11-11     0    57
#>  3 Melbourne Central                 2019-11-11 00:00:00 2019-11-11     0   235
#>  4 Town Hall (West)                  2019-11-11 00:00:00 2019-11-11     0   129
#>  5 Princes Bridge                    2019-11-11 00:00:00 2019-11-11     0    76
#>  6 Flinders Street Station Underpass 2019-11-11 00:00:00 2019-11-11     0   151
#>  7 Birrarung Marr                    2019-11-11 00:00:00 2019-11-11     0     4
#>  8 Webb Bridge                       2019-11-11 00:00:00 2019-11-11     0    15
#>  9 Southern Cross Station            2019-11-11 00:00:00 2019-11-11     0    20
#> 10 Victoria Point                    2019-11-11 00:00:00 2019-11-11     0     8
#> # … with 7,214 more rows

# Retrieve data of a speficied period
start_date <- as.Date("2017-07-01")
end_date <- start_date + 6L
melb_walk(from = start_date, to = end_date)

#> # A tibble: 7,224 x 5
#>    Sensor                            Date_Time           Date        Time Count
#>    <chr>                             <dttm>              <date>     <int> <int>
#>  1 Bourke Street Mall (North)        2017-07-01 00:00:00 2017-07-01     0   280
#>  2 Bourke Street Mall (South)        2017-07-01 00:00:00 2017-07-01     0   177
#>  3 Melbourne Central                 2017-07-01 00:00:00 2017-07-01     0   826
#>  4 Town Hall (West)                  2017-07-01 00:00:00 2017-07-01     0   682
#>  5 Princes Bridge                    2017-07-01 00:00:00 2017-07-01     0     0
#>  6 Flinders Street Station Underpass 2017-07-01 00:00:00 2017-07-01     0   768
#>  7 Birrarung Marr                    2017-07-01 00:00:00 2017-07-01     0    74
#>  8 Webb Bridge                       2017-07-01 00:00:00 2017-07-01     0    22
#>  9 Southern Cross Station            2017-07-01 00:00:00 2017-07-01     0    35
#> 10 Victoria Point                    2017-07-01 00:00:00 2017-07-01     0    27
#> # … with 7,214 more rows

# }