API using compedapi to Melbourne pedestrian data

Provides API using compedapi to Melbourne pedestrian data in a tidy data form.

melb_walk(from = to - 6L, to = Sys.Date() - 1L, na.rm = FALSE, session = NULL)

Arguments

from
Starting date.

to
Ending date.

na.rm
Logical. FALSE is the default suggesting to include NA in the dataset. TRUE removes the NAs.

session
NULL or "shiny". For internal use only.

Value

A tibble including these variables as follows:

Sensor: Sensor name (43 sensors up to date)

Date_Time: Date time when the pedestrian counts are recorded

Date: Date associated with Date_Time

Time: Time of day

Count: Hourly counts

Details

It provides API using compedapi, where counts are uploaded on a daily basis. The up-to-date data would be till the previous day. The data is sourced from Melbourne Open Data Portal. Please refer to Melbourne Open Data Portal for more details about the dataset and its policy.

See also

Examples

if (FALSE) {
# Retrieve last week data
melb_walk()

# Retrieve data of a speficied period
start_date <- as.Date("2017-07-01")
end_date <- start_date + 6L
melb_walk(from = start_date, to = end_date)
}