API using Socrata to Melbourne pedestrian data (per hour)

API using Socrata to Melbourne pedestrian data (per hour)

melb_walk_fast(year = NULL, sensor = NULL, na.rm = FALSE, app_token = NULL)

Arguments

year
An integer or a vector of integers. By default, it's the current year.

sensor
Sensor names. By default, it pulls all the sensors. Use pull_sensor to see the available sensors.

na.rm
Logical. FALSE is the default suggesting to include NA in the dataset. TRUE removes the NAs.

app_token
Characters giving the application token. A limited number of requests can be made without an app token (NULL), but they are subject to much lower throttling limits than request that do include one. Sign up for an app token here.

Value

A tibble including these variables as follows:

Sensor: Sensor name

Date_Time: Date time when the pedestrian counts are recorded

Date: Date associated with date_Time

Time: Time of day

Count: Hourly counts

Details

It provides the API using Socrata, where counts are uploaded on a monthly basis. The up-to-date data would be till the previous month. The data is sourced from Melbourne Open Data Portal. Please refer to Melbourne Open Data Portal for more details about the dataset and its policy.

See also

Examples

# \dontrun{
# Retrieve the year 2017
melb_walk_fast(year = 2017)

#> # A tibble: 455,520 x 5
#>    Sensor                         Date_Time           Date        Time Count
#>    <chr>                          <dttm>              <date>     <int> <int>
#>  1 Alfred Place                   2017-01-01 00:00:00 2017-01-01     0   222
#>  2 Birrarung Marr                 2017-01-01 00:00:00 2017-01-01     0  1099
#>  3 Bourke St-Russell St (West)    2017-01-01 00:00:00 2017-01-01     0  1916
#>  4 Bourke Street Mall (North)     2017-01-01 00:00:00 2017-01-01     0   855
#>  5 Bourke Street Mall (South)     2017-01-01 00:00:00 2017-01-01     0   855
#>  6 Chinatown-Lt Bourke St (South) 2017-01-01 00:00:00 2017-01-01     0   353
#>  7 Chinatown-Swanston St (North)  2017-01-01 00:00:00 2017-01-01     0   697
#>  8 City Square                    2017-01-01 00:00:00 2017-01-01     0  5120
#>  9 Collins Place (North)          2017-01-01 00:00:00 2017-01-01     0   444
#> 10 Collins Place (South)          2017-01-01 00:00:00 2017-01-01     0   844
#> # … with 455,510 more rows

# Retrieve the year 2017 for Southern Cross Station
melb_walk_fast(year = 2017, sensor = "Southern Cross Station")

#> # A tibble: 8,760 x 5
#>    Sensor                 Date_Time           Date        Time Count
#>    <chr>                  <dttm>              <date>     <int> <int>
#>  1 Southern Cross Station 2017-01-01 00:00:00 2017-01-01     0  1335
#>  2 Southern Cross Station 2017-01-01 01:00:00 2017-01-01     1   463
#>  3 Southern Cross Station 2017-01-01 02:00:00 2017-01-01     2   219
#>  4 Southern Cross Station 2017-01-01 03:00:00 2017-01-01     3   122
#>  5 Southern Cross Station 2017-01-01 04:00:00 2017-01-01     4    21
#>  6 Southern Cross Station 2017-01-01 05:00:00 2017-01-01     5    28
#>  7 Southern Cross Station 2017-01-01 06:00:00 2017-01-01     6    26
#>  8 Southern Cross Station 2017-01-01 07:00:00 2017-01-01     7    35
#>  9 Southern Cross Station 2017-01-01 08:00:00 2017-01-01     8    55
#> 10 Southern Cross Station 2017-01-01 09:00:00 2017-01-01     9    82
#> # … with 8,750 more rows

# }