API using Socrata to Melbourne pedestrian sensor locations

Provides API using Socrata to Melbourne pedestrian sensor locations.

pull_sensor(app_token = NULL)

Arguments

app_token
Characters giving the application token. A limited number of requests can be made without an app token (NULL), but they are subject to much lower throttling limits than request that do include one. Sign up for an app token here.

Details

It provides API using Socrata.

See also

Examples

# \dontrun{
pull_sensor()

#> # A tibble: 65 x 9
#>    sensor sensor_id longitude latitude direction_1 direction_2
#>    <chr>  <chr>         <dbl>    <dbl> <chr>       <chr>      
#>  1 Build… 59             145.    -37.8 North       South      
#>  2 Spenc… 23             145.    -37.8 East        West       
#>  3 China… 20             145.    -37.8 East        West       
#>  4 Flind… 34             145.    -37.8 East        West       
#>  5 Bourk… 57             145.    -37.8 West        East       
#>  6 Lonsd… 40             145.    -37.8 South       North      
#>  7 Lygon… 31             145.    -37.8 South       North      
#>  8 China… 19             145.    -37.8 East        West       
#>  9 Lygon… 37             145.    -37.8 North       South      
#> 10 Queen… 36             145.    -37.8 South       North      
#> # … with 55 more rows, and 3 more variables: installation_date <dttm>,
#> #   status <chr>, note <chr>

# }