API using Socrata to Melbourne pedestrian sensor locations

Provides API using Socrata to Melbourne pedestrian sensor locations.

pull_sensor(app_token = NULL)

Arguments

app_token
Characters giving the application token. A limited number of requests can be made without an app token (NULL), but they are subject to much lower throttling limits than request that do include one. Sign up for an app token here.

Value

A data frame including these variables as follows:

Sensor: Sensor name (43 sensors up to date)

Sensor_ID: Sensor identifier

Longitude: Longitude

Latitude: Latitude

Location_Type: Location type

Year_Installed: Year installed

Details

It provides API using Socrata.

See also

run_melb

Examples

# NOT RUN {
  pull_sensor()
# }