API using Socrata to Melbourne pedestrian sensor locations

Provides API using Socrata to Melbourne pedestrian sensor locations.

pull_sensor(app_token = NULL)

Arguments

app_token
Characters giving the application token. A limited number of requests can be made without an app token (NULL), but they are subject to much lower throttling limits than request that do include one. Sign up for an app token here.

Details

It provides API using Socrata.

See also

Examples

# \dontrun{
pull_sensor()

#>                                                                        sensor
#> 1                                                            Building 80 RMIT
#> 2                                               Spencer St-Collins St (South)
#> 3                                              Chinatown-Lt Bourke St (South)
#> 4                                                        Flinders St-Spark La
#> 5                                                            Bourke St Bridge
#> 6                                                Lonsdale St-Spring St (West)
#> 7                                                             Lygon St (West)
#> 8                                               Chinatown-Swanston St (North)
#> 9                                                             Lygon St (East)
#> 10                                                            Queen St (West)
#> 11                                             Bourke St - Spencer St (North)
#> 12                                                                Webb Bridge
#> 13                                                            Waterfront City
#> 14                                              St Kilda Rd-Alexandra Gardens
#> 15                                                      Collins Place (North)
#> 16                                                      QVM-Therry St (South)
#> 17                                              QV Market-Elizabeth St (West)
#> 18                                                       Australia on Collins
#> 19                                            Flinders St-Elizabeth St (East)
#> 20                                                      Collins Place (South)
#> 21                                              Spencer St-Collins St (North)
#> 22                                                               Alfred Place
#> 23                                                          QV Market-Peel St
#> 24                                                              State Library
#> 25                                                            The Arts Centre
#> 26                                               Tin Alley-Swanston St (West)
#> 27 Melbourne Central-Elizabeth St (East)Melbourne Central-Elizabeth St (East)
#> 28                                                        QVM-Queen St (East)
#> 29                                                    QVM-Franklin St (North)
#> 30                                                             Victoria Point
#> 31                                               Monash Rd-Swanston St (West)
#> 32                                                                   New Quay
#> 33                                                     Collins Street (North)
#> 34                                                          Melbourne Central
#> 35                                                     Southern Cross Station
#> 36                                                    Lincoln-Swanston (West)
#> 37                                                        Lonsdale St (South)
#> 38                                          Flinders Street Station Underpass
#> 39                                             Flinders St-Swanston St (West)
#> 40                                                                City Square
#> 41                                                Bourke St-Russell St (West)
#> 42                                                           Town Hall (West)
#> 43                                                          Pelham St (South)
#> 44                                         Lonsdale St - Elizabeth St (North)
#> 45                                                          Flagstaff Station
#> 46                                              Grattan St-Swanston St (West)
#> 47                                                             Princes Bridge
#> 48                                                 Faraday St-Lygon St (West)
#> 49                                           Elizabeth St-Lonsdale St (South)
#> 50                                                 Bourke Street Mall (South)
#> 51                                                                  Southbank
#> 52                                                 Bourke Street Mall (North)
#> 53                                       Little Collins St-Swanston St (East)
#> 54                                                           RMIT Building 14
#> 55                                                             Birrarung Marr
#> 56                                            Elizabeth St-La Trobe St (East)
#> 57                                                        La Trobe St (North)
#> 58                                               Flinders St-Spring St (West)
#> 59                                                           Sandridge Bridge
#> 60                                             Flinders La-Swanston St (West)
#>    sensor_id longitude  latitude direction_1 direction_2 installation_date
#> 1         59  144.9630 -37.80826       North       South        2019-02-13
#> 2         23  144.9545 -37.81909        East        West        2013-09-02
#> 3         20  144.9682 -37.81173        East        West        2013-09-06
#> 4         34  144.9742 -37.81538        East        West        2014-06-08
#> 5         57  144.9503 -37.81767        West        East        2018-08-13
#> 6         40  144.9723 -37.80999       South       North        2015-01-19
#> 7         31  144.9666 -37.80170       South       North        2013-10-10
#> 8         19  144.9655 -37.81237        East        West        2013-09-02
#> 9         37  144.9667 -37.80310       North       South        2015-02-11
#> 10        36  144.9612 -37.81653       South       North        2015-01-20
#> 11        58  144.9536 -37.81686        West        East        2018-09-27
#> 12         8  144.9472 -37.82294       North       South        2009-03-24
#> 13        11  144.9397 -37.81565        East        West        2009-01-20
#> 14        29  144.9686 -37.81966       South       North        2013-10-11
#> 15        18  144.9731 -37.81345        East        West        2009-03-30
#> 16        49  144.9596 -37.80730        West        East        2017-11-29
#> 17        26  144.9605 -37.80870       South       North        2013-09-28
#> 18        16  144.9652 -37.81573        <NA>        <NA>        2009-03-30
#> 19        22  144.9651 -37.81786       South       North        2013-08-12
#> 20        17  144.9732 -37.81363        East        West        2009-03-30
#> 21        24  144.9545 -37.81888        East        West        2013-09-02
#> 22        39  144.9700 -37.81380       South       North        2014-08-27
#> 23        27  144.9564 -37.80607        East        West        2013-08-16
#> 24        15  144.9645 -37.81064       South       North        2009-03-25
#> 25        28  144.9688 -37.82130       South       North        2013-08-23
#> 26        44  144.9644 -37.79690       South       North        2015-04-15
#> 27        47  144.9622 -37.81160       South       North        2017-08-24
#> 28        48  144.9587 -37.80632        East        West        2017-10-02
#> 29        51  144.9591 -37.80842        West        East        2017-11-30
#> 30        10  144.9471 -37.81876        West        East        2009-04-23
#> 31        43  144.9641 -37.79845       South       North        2015-04-15
#> 32        12  144.9429 -37.81458        West        East        2009-01-21
#> 33        53  144.9655 -37.81564        West        East        2015-09-23
#> 34         3  144.9643 -37.81102       South       North        2009-03-25
#> 35         9  144.9510 -37.81983        East        West        2009-03-23
#> 36        54  144.9631 -37.80402       South       North        2018-06-26
#> 37        30  144.9666 -37.81122        East        West        2013-10-14
#> 38         6  144.9656 -37.81912       North       South        2009-03-25
#> 39        38  144.9672 -37.81723        <NA>        <NA>        2014-12-05
#> 40        32  144.9669 -37.81574        <NA>        <NA>        2013-12-20
#> 41        21  144.9678 -37.81245       South       North        2013-09-02
#> 42         4  144.9661 -37.81488       South       North        2009-03-23
#> 43        46  144.9616 -37.80241        West        East        2017-07-10
#> 44        56  144.9615 -37.81235        East        West        2018-07-25
#> 45        13  144.9565 -37.81240        <NA>        <NA>        2009-03-24
#> 46        42  144.9639 -37.80009       South       North        2015-04-15
#> 47         5  144.9679 -37.81874       North       South        2009-03-26
#> 48        50  144.9672 -37.79808       North       South        2017-11-30
#> 49        52  144.9619 -37.81252        West        East        2017-07-31
#> 50         2  144.9652 -37.81381        East        West        2009-03-30
#> 51        35  144.9651 -37.82018        East        West        2016-04-11
#> 52         1  144.9652 -37.81349        West        East        2009-03-24
#> 53        45  144.9661 -37.81414       North       South        2017-06-29
#> 54        61  144.9631 -37.80767       North       South        2019-06-28
#> 55         7  144.9717 -37.81863        East        West        2014-12-17
#> 56        55  144.9613 -37.80989       North       South        2018-07-19
#> 57        62  144.9622 -37.80996        East        West        2019-09-25
#> 58        33  144.9745 -37.81483       North       South        2014-04-23
#> 59        14  144.9629 -37.82011       North       South        2019-09-25
#> 60        41  144.9669 -37.81669       North       South        2017-06-29
#>    status                                        note
#> 1       A                                        <NA>
#> 2       A                                        <NA>
#> 3       A                                        <NA>
#> 4       A                                        <NA>
#> 5       A                                        <NA>
#> 6       A                                        <NA>
#> 7       A                                        <NA>
#> 8       A                                        <NA>
#> 9       A                                        <NA>
#> 10      A                                        <NA>
#> 11      A                                        <NA>
#> 12      A                                        <NA>
#> 13      A                                        <NA>
#> 14      A                                        <NA>
#> 15      A                                        <NA>
#> 16      A                                        <NA>
#> 17      A                                        <NA>
#> 18      R Device moved to location ID 53 (22/09/2015)
#> 19      A                                        <NA>
#> 20      A                                        <NA>
#> 21      A                                        <NA>
#> 22      A                                        <NA>
#> 23      A                                        <NA>
#> 24      A                                        <NA>
#> 25      A                                        <NA>
#> 26      A                                        <NA>
#> 27      A                                        <NA>
#> 28      A                                        <NA>
#> 29      A                                        <NA>
#> 30      A                                        <NA>
#> 31      A                                        <NA>
#> 32      A                                        <NA>
#> 33      A                                        <NA>
#> 34      A                                        <NA>
#> 35      A                                        <NA>
#> 36      A                                        <NA>
#> 37      A                                        <NA>
#> 38      A                                        <NA>
#> 39      R        Device has been removed (17/02/2017)
#> 40      R        Device has been removed (24/01/2017)
#> 41      A                                        <NA>
#> 42      A                                        <NA>
#> 43      A                                        <NA>
#> 44      A                                        <NA>
#> 45      I                                        <NA>
#> 46      A                                        <NA>
#> 47      A                                        <NA>
#> 48      A                                        <NA>
#> 49      A                                        <NA>
#> 50      A                                        <NA>
#> 51      A                                        <NA>
#> 52      A                                        <NA>
#> 53      A                                        <NA>
#> 54      A                                        <NA>
#> 55      A                                        <NA>
#> 56      A                                        <NA>
#> 57      A                                        <NA>
#> 58      I                                        <NA>
#> 59      A                                        <NA>
#> 60      A                                        <NA>

# }